ARM: dts: msm: Limit ADSP CMA heap to lowmem for 8x26
Commit 8360c416dc5bdabbfcc22a209eb9cc6c47be99ac (Remove fb
memory allocation) freed 8MB allocation from highmem which
resulted in ADSP CMA heap being allocated in highmem instead
of lowmem. This allocation along with other CMA heap allocations
from highmem resulted in 'totalhigh_pages' to be calculated as
zero at bootup and entire highmem being intialized with
migrate_type as MIGRATE_CMA for 8626. This lead to performance
degradation for some benchmark tests.

To resolve this, limit the ADSP CMA heap to lowmem so that the
'totalhigh_pages' are not zero at bootup and not all highmem
is initialized to MIGRATE_CMA.

Change-Id: I55f49674696c81ee35944335d9d73d6f9fefa20c
Signed-off-by: Nirmal Abraham <nabrah@codeaurora.org>